+++ src/main/java/com/lucianaugusto/ebanxassignment/event/http/request/EventHttpRequest.java
@@ -2,9 +2,7 @@ package com.lucianaugusto.ebanxassignment.event.http.request;
 
 import com.lucianaugusto.ebanxassignment.event.http.enums.EventTypeEnum;
 
-import java.math.BigDecimal;
-
-public record EventHttpRequest(EventTypeEnum type, String origin, String destination, BigDecimal amount) {
+public record EventHttpRequest(EventTypeEnum type, String origin, String destination, Integer amount) {
     public boolean validate() {
         if (type == null) {
             return false;
