+++ src/main/java/com/lucianaugusto/ebanxassignment/event/http/response/EventAccountInfo.java
@@ -1,5 +1,3 @@
 package com.lucianaugusto.ebanxassignment.event.http.response;
 
-import java.math.BigDecimal;
-
-public record EventAccountInfo(String id, BigDecimal amount) {}
+public record EventAccountInfo(String id, Integer amount) {}
